requirements.txt
argparse
python-dotenv
networkx
matplotlib
numpy
seaborn